+++ src/events/MicrosoftDynamics365TargetSchemasEvent.php
@@ -1,0 +1,13 @@
+<?php
+namespace verbb\formie\events;
+
+use yii\base\Event;
+
+class MicrosoftDynamics365TargetSchemasEvent extends Event
+{
+    // Properties
+    // =========================================================================
+
+    public array $targetSchemas;
+
+}