+++ src/jobs/TriggerIntegration.php
@@ -27,7 +27,7 @@ class TriggerIntegration extends BaseJob
     /**
      * @inheritDoc
      */
-    public function getDescription(): string
+    public function execute($queue)
     {
         $formName = '';
 
@@ -84,5 +84,16 @@ class TriggerIntegration extends BaseJob
         }
 
         $this->setProgress($queue, 1);
+    }
+
+    // Protected Methods
+    // =========================================================================
+
+    /**
+     * @inheritDoc
+     */
+    protected function defaultDescription(): string
+    {
+        return Craft::t('formie', 'Triggering form “{handle}” integration.', ['handle' => $this->integration->handle]);
     }
 }